feat: move edit book page to modal
