+++ src/components/admin/work-item.tsx
@@ -1,11 +1,11 @@
 "use client";
-import Link from "next/link";
 import { type Book } from "@prisma/client";
 
 import { cn } from "@/lib/utils";
 
 import { buttonVariants } from "../ui/button";
 import { DeleteButton } from "./delete-button";
+import { EditBook } from "./edit";
 
 export const WorkItem = ({ book }: { book: Book }) => {
   return (
@@ -21,20 +21,7 @@ export const WorkItem = ({ book }: { book: Book }) => {
       {/* actions */}
       <div className="flex items-center justify-between">
         <div className="space-x-3">
-          <Link
-            href={{
-              pathname: `/admin/books/edit/${book.id}`,
-              query: {
-                title: book.title,
-                grade: book.grade,
-                language: book.language,
-                term: book.term,
-              },
-            }}
-            className={cn(buttonVariants({ size: "sm" }))}
-          >
-            Өзгерту
-          </Link>
+          <EditBook book={book} />
 
           <DeleteButton book={book} />
         </div>
